Avoid some more clones
